Cleanup & user stuff.
In the previous commit, I'd thought that it'd be unnecessary to really
ever add keys to one of these machines; then I realized that for VBox
VMs, I'm as likely to be using them directly as through a remote session
(e.g. ssh -A), so may want to have new keys on them (and therefore in
Github.) GCE probably not so much, but whatever.

A bit of refactoring in u/me/prompt, and more in ambroix.